+++ src/rust/ensogl/src/data.rs
@@ -7,4 +7,5 @@ pub mod function;
 pub mod seq;
 pub mod theme;
 
+pub use data::hash_map_tree::HashMapTree;
 pub use data::opt_vec::OptVec;
