+++ src/rust/ensogl/src/display.rs
@@ -10,8 +10,6 @@ pub mod scene;
 pub mod shape;
 pub mod symbol;
 pub mod style;
-pub mod style2;
-pub mod style3;
 pub mod world;
