+++ src/rust/lib/debug-scenes/src/shapes.rs
@@ -13,9 +13,7 @@ use graph_editor::GraphEditor;
 use wasm_bindgen::prelude::*;
 use ensogl::display::object::ObjectOps;
 
-//use ensogl::display::style;
-//use ensogl::display::style2;
-use ensogl::display::style3;
+use ensogl::display::style;
 
 
 #[wasm_bindgen]
@@ -58,7 +56,5 @@ fn init(app:&Application) {
         was_rendered = true;
     }).forget();
 
-//    style::test();
-//    style2::test();
-    style3::test();
+    style::test();
 }
